+++ api/app/model/examengeneral_model.php
@@ -2,15 +2,14 @@
 
 namespace App\Model;
 
-use App\Lib\Response,
-	App\Lib\Security;
-
+use App\Lib\Response;
 /**
 * Modelo examen general
 */
 class  ExamenGeneralModel
 {
 	private $db;
+	private $db_pdo;
 	private $table = 'examen_general';
 	private $response;
 
@@ -20,19 +19,17 @@ class  ExamenGeneralModel
 		$this->db 		= $db;
 		$this->db_pdo   = $db_pdo;
 		$this->response = new Response();
-		$this->security = new Security();
 	}
-
-	//var $l => 'limit', $p => 'pagina'
 
 	//lista_total
 	public function listar(){
-
-		return $data = $this->db->from($this->table)
+		$data = $this->db->from($this->table)
 						 ->orderBy('id DESC')
 						 ->fetchAll();
-	//  return $data = $this->db_pdo->query('select * from '.$this->table)
-	//					 			->fetchAll();
+		if ($data != null){
+			return $this->response->setResponse(true, $data, '0');
+		}
+		return $this->response->setResponse(true, 'No existen datos', '1');
 	}
 
 	//listar paginado
@@ -59,24 +56,15 @@ class  ExamenGeneralModel
 	//obtener
 	public function getExamenGeneral($id){
 
-		return $data = $this->db->from($this->table, $id)
+		$data = $this->db->from($this->table, $id)
 								->fetch();
+		if ($data != null){
+			return $this->response->setResponse(true, $data, '0');
+		}
+		return $this->response->setResponse(true, 'Registro no encontrado', '1');
 	}
 	//registrar
-
-	public function insert($data){
-		// $data['password'] = md5($data['password']);
-
-		$this->db_pdo->insertInto($this->table, $data)
-				 ->execute();
-
-		return $this->response->setResponse(true);
-		}
-
 	public function insertGeneralTest($data){
-
-		//$this->db->insertInto($this->table, $data)
-		//		 ->execute();
 		$this->db_pdo->multi_query(" CALL insertarExamenGeneral('".$data['_color']."',
 														'".$data['_cantidad']."',
 														'".$data['_olor']."',
@@ -112,18 +100,29 @@ class  ExamenGeneralModel
 
 	//actualizar
 	public function update($data, $id){
+		$oldData = $this->db->from($this->table, $id)
+		->fetch();
 
-		$this->db->update($this->table, $data, $id)
-				 ->execute();
+		if ($oldData != null) {
+			$this->db->update($this->table, $data, $id)
+			->execute();
 
-		return $this->response->setResponse(true);
+   			return $this->response->setResponse(true, 'El registro se actualizó correctamente', '0');
+		}
+		return $this->response->setResponse(true, 'Error al actualizar, el registro no existe', '1');
 	}
 	//eliminar
 	public function delete($id){
-		$this->db->deleteFrom($this->table, $id)
+		$data = $this->db->from($this->table, $id)
+		->fetch();
+
+		if ($data != null) {
+			$this->db->deleteFrom($this->table, $id)
 			->execute();
 
-		return $this->response->setResponse(true);
+			return $this->response->setResponse(true, 'Registro eliminado', '0');
+		}
+		return $this->response->setResponse(true, 'Error al eliminar, el registro no existe', '1');
 	}
 
 }
