+++ api/app/model/informes_g.php
@@ -2,8 +2,7 @@
 
 namespace App\Model;
 
-use App\Lib\Response,
-	App\Lib\Security;
+use App\Lib\Response;
 
 /**
 * Modelo usuario
@@ -11,6 +10,7 @@ use App\Lib\Response,
 class  InformesGModel
 {
 	private $db;
+	private $db_pdo;
 	private $table = 'informes_g';
 	private $response;
 
@@ -20,19 +20,17 @@ class  InformesGModel
 		$this->db 		= $db;
 		$this->db_pdo   = $db_pdo;
 		$this->response = new Response();
-		$this->security = new Security();
 	}
-
-	//var $l => 'limit', $p => 'pagina'
 
 	//lista_total
 	public function listar(){
-
-		return $data = $this->db->from($this->table)
+		$data = $this->db->from($this->table)
 						 ->orderBy('id DESC')
 						 ->fetchAll();
-	//  return $data = $this->db_pdo->query('select * from '.$this->table)
-	//					 			->fetchAll();
+		if ($data != null){
+			return $this->response->setResponse(true, $data, '0');
+		}
+		return $this->response->setResponse(true, 'No existen datos', '1');
 	}
 
 	//listar paginado
@@ -58,21 +56,14 @@ class  InformesGModel
 	}
 	//obtener
 	public function getInformeG($id){
-
-		return $data = $this->db->from($this->table, $id)
+		$data = $this->db->from($this->table, $id)
 								->fetch();
+		if ($data != null){
+			return $this->response->setResponse(true, $data, '0');
+		}
+		return $this->response->setResponse(true, 'Registro no encontrado', '1');
 	}
 	//registrar
-
-	public function insert($data){
-		// $data['password'] = md5($data['password']);
-
-		$this->db_pdo->insertInto($this->table, $data)
-				 ->execute();
-
-		return $this->response->setResponse(true);
-		}
-
 	public function insertarInformeG($data){
 		$this->db_pdo->multi_query(" CALL insertarInformeG('".$data['_nombre']."',
 														'".$data['_contenido']."',
@@ -87,18 +78,29 @@ class  InformesGModel
 	}
 	//actualizar
 	public function update($data, $id){
+		$oldData = $this->db->from($this->table, $id)
+		->fetch();
 
-		$this->db->update($this->table, $data, $id)
-				 ->execute();
-
-		return $this->response->setResponse(true);
-	}
-	//eliminar
-	public function delete($id,$titulo){
-		$this->db->deleteFrom($this->table, $id)
+		if ($oldData != null) {
+			$this->db->update($this->table, $data, $id)
 			->execute();
 
-		return $this->response->setResponse(true);
+   			return $this->response->setResponse(true, 'El registro se actualizó correctamente', '0');
+		}
+		return $this->response->setResponse(true, 'Error al actualizar, el registro no existe', '1');
+	}
+	//eliminar
+	public function delete($id){
+		$data = $this->db->from($this->table, $id)
+		->fetch();
+
+		if ($data != null) {
+			$this->db->deleteFrom($this->table, $id)
+			->execute();
+
+			return $this->response->setResponse(true, 'Registro eliminado', '0');
+		}
+		return $this->response->setResponse(true, 'Error al eliminar, el registro no existe', '1');
 	}
 
 }
