+++ api/app/model/medico_model.php
@@ -2,8 +2,7 @@
 
 namespace App\Model;
 
-use App\Lib\Response,
-	App\Lib\Security;
+use App\Lib\Response;
 
 /**
 * Modelo usuario
@@ -11,6 +10,7 @@ use App\Lib\Response,
 class  MedicoModel
 {
 	private $db;
+	private $db_pdo;
 	private $table = 'medico';
 	private $response;
 
@@ -20,23 +20,18 @@ class  MedicoModel
 		$this->db 		= $db;
 		$this->db_pdo   = $db_pdo;
 		$this->response = new Response();
-		$this->security = new Security();
 	}
-
-	//var $l => 'limit', $p => 'pagina'
 
 	//lista_total
 	public function listar(){
-
-				$data = $this->db->from($this->table)
+		$data = $this->db->from($this->table)
 						 ->orderBy('id DESC')
 						 ->fetchAll();
-				if(empty($data)){
-		 			$respuesta = array('respuesta' => 0 );
-		 			return $respuesta;
-				}else{
-		 			return $data;
-				}
+
+		if ($data != null){
+			return $this->response->setResponse(true, $data, '0');
+		}
+		return $this->response->setResponse(true, 'No existen datos', '1');
 	}
 
 	//listar paginado
@@ -62,35 +57,48 @@ class  MedicoModel
 	}
 	//obtener
 	public function getMedico($id){
-
-		return $data = $this->db->from($this->table, $id)
+		$data = $this->db->from($this->table, $id)
 								->fetch();
+		if ($data != null){
+			return $this->response->setResponse(true, $data, '0');
+		}
+		return $this->response->setResponse(true, 'Registro no encontrado', '1');
 	}
 	//registrar
 
 	public function insert($data){
-		// $data['password'] = md5($data['password']);
-
-		$this->db->insertInto($this->table, $data)
-				 ->execute();
-
-		return $this->response->setResponse(true);
-		}
+		$this->db_pdo->multi_query(" CALL insertarDoctor('".$data['_nombre']."',
+													'".$data['_apellidos']."')");
+			$res = $this->db_pdo->store_result();
+			$res = $res->fetch_assoc();
+			mysqli_close($this->db_pdo);
+			return $this->response->setResponse(true, $res['respuesta'], $res['error']);
+	}
 	//actualizar
 	public function update($data, $id){
+		$oldData = $this->db->from($this->table, $id)
+		->fetch();
 
-		$this->db->update($this->table, $data, $id)
-				 ->execute();
+		if ($oldData != null) {
+			$this->db->update($this->table, $data, $id)
+			->execute();
 
-		return $this->response->setResponse(true);
+   			return $this->response->setResponse(true, 'El registro se actualizó correctamente', '0');
+		}
+		return $this->response->setResponse(true, 'Error al actualizar, el registro no existe', '1');
 	}
 	//eliminar
 	public function delete($id){
+		$data = $this->db->from($this->table, $id)
+		->fetch();
 
-		$this->db->deleteFrom($this->table, $id)
-				 ->execute();
+		if ($data != null) {
+			$this->db->deleteFrom($this->table, $id)
+			->execute();
 
-		return $this->response->setResponse(true);
+			return $this->response->setResponse(true, 'Registro eliminado', '0');
+		}
+		return $this->response->setResponse(true, 'Error al eliminar, el registro no existe', '1');
 	}
 
 }
